chatbot-core/requirements.txt
# =========================
# Testing
# =========================
pytest==8.4.1
pytest-cov==6.2.1
pytest-mock==3.14.1

# =========================
# Core API Dependencies
# =========================
fastapi==0.115.12
pydantic==2.12.5
uvicorn[standard]==0.34.3
starlette==0.46.2

# =========================
# LangChain & AI
# =========================
langchain==0.3.25
langchain-core==0.3.61
langchain-text-splitters==0.3.8
langsmith==0.3.42

# =========================
# ML & Transformers
# =========================
sentence-transformers==4.1.0
transformers==4.54.1
tokenizers==0.21.1

# Optional LLM binding (skip if build toolchain unavailable)
# llama-cpp-python==0.3.9

# =========================
# PyTorch (GPU)
# =========================
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
triton==3.3.1; sys_platform == 'linux'

# =========================
# NVIDIA CUDA (GPU-only, Python <3.13)
# =========================
nvidia-cublas-cu12==12.6.4.1; sys_platform != 'darwin'
nvidia-cuda-cupti-cu12==12.6.80; sys_platform != 'darwin'
nvidia-cuda-nvrtc-cu12==12.6.77; sys_platform != 'darwin'
nvidia-cuda-runtime-cu12==12.6.77; sys_platform != 'darwin'
nvidia-cudnn-cu12==9.5.1.17; sys_platform != 'darwin'
nvidia-cufft-cu12==11.3.0.4; sys_platform != 'darwin'
nvidia-cufile-cu12==1.11.1.6; python_version < "3.13" and sys_platform != 'darwin'
nvidia-curand-cu12==10.3.7.77; sys_platform != 'darwin'
nvidia-cusolver-cu12==11.7.1.2; sys_platform != 'darwin'
nvidia-cusparse-cu12==12.5.4.2; sys_platform != 'darwin'
nvidia-cusparselt-cu12==0.6.3; sys_platform != 'darwin'
nvidia-nccl-cu12==2.26.2; sys_platform != 'darwin'
nvidia-nvjitlink-cu12==12.6.85; sys_platform != 'darwin'
nvidia-nvtx-cu12==12.6.77; sys_platform != 'darwin'

# =========================
# Vector Store & Retrieval
# =========================
faiss-cpu==1.13.2

# Optional retriever extras; comment out if build tools unavailable
retriv @ git+https://github.com/giovanni-vaccarino/retriv.git@remove-pystemmer

# =========================
# Data Processing
# =========================
numpy==2.2.6
pandas==2.2.3
scikit-learn==1.6.1
scipy==1.15.3

# Numba JIT (required by vendored retriv dense retriever)
numba==0.63.1
llvmlite==0.46.0

# =========================
# Database
# =========================
SQLAlchemy==2.0.41

# =========================
# Web & HTTP
# =========================
httpx==0.28.1
httpcore==1.0.9
httptools==0.6.4
requests==2.32.3
requests-toolbelt==1.0.0

# =========================
# File Processing (from main branch)
# =========================
python-magic==0.4.27
python-multipart==0.0.20

# =========================
# Utilities
# =========================
python-dotenv==1.1.0
PyYAML==6.0.2
tqdm==4.67.1
click==8.2.1
python-dateutil==2.9.0.post0
setuptools==80.9.0
typing-extensions>=4.14.1
typing-inspection>=0.4.2
diskcache==5.6.3
filelock==3.18.0
fsspec==2025.5.1
greenlet==3.2.2
h11==0.16.0
hf-xet==1.1.5
Jinja2==3.1.6
MarkupSafe==3.0.2
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
orjson==3.10.18
packaging==24.2
pillow==11.2.1
platformdirs==4.3.8
pyparsing==3.2.3
pytz==2025.2
regex==2024.11.6
safetensors==0.5.3
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tomlkit==0.13.2
tzdata==2025.2
zstandard==0.23.0
annotated-types==0.7.0
anyio==4.9.0
astroid==3.3.10
contourpy==1.3.2
cycler==0.12.1
dill==0.4.0
mccabe==0.7.0
mpmath==1.3.0
networkx==3.4.2
psutil==7.1.3

# =========================
# Code Quality
# =========================
pylint==3.3.7
isort==6.0.1

# =========================
# Additional retriv dependencies
# =========================
oneliner_utils==0.1.2
indxr==0.1.6
multipipe==0.1.0
nltk==3.9.2
optuna==4.6.0
ranx==0.3.21
Unidecode==1.4.0
